Fix cast_column_menu.ts

+++ ui/src/components/widgets/sql/table/menus/cast_column_menu.ts
@@ -33,8 +33,8 @@ const CASTS = {
   int: {
     type: PerfettoSqlTypes.INT,
   },
-  float: {
-    type: PerfettoSqlTypes.FLOAT,
+  double: {
+    type: PerfettoSqlTypes.DOUBLE,
   },
   string: {
     type: PerfettoSqlTypes.STRING,
